Make a base tag class for better managing position and size.
