+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/LyricRows/Lyrics/Blueprints/RubyTags/RubySelectionBlueprint.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Game.Rulesets.Karaoke.Objects.Types;
+
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.LyricRows.Lyrics.Blueprints.RubyTags
+{
+    public class RubySelectionBlueprint : TagSelectionBlueprint<ITextTag>
+    {
+        public RubySelectionBlueprint(ITextTag item)
+            : base(item)
+        {
+        }
+    }
+}
